Add files via upload
This was the first project I tried on my own that includes pictures, multiple links and even features a contact page. I had to use the MDN web docs page to find out how to attach a file to the Contact Us page. I began this project on 4/26/22. It is still a work in progress. I'm learning a lot along the way. Not sure if I will add to CSS to this project.